Implement OUT node
I guess this isn't just a NOP, it's actually a MOV.  The remainder of
the C++ wrapper handles copying code to the output so it can actually
be used, so there's not really anything we need to do.  This MOV
should get completely elided by the optimizer, so it's really just a
NOP :).