Use libflo's node::outwid() instead of width()
This is the correct thing to use when generating output node widths,
but it's not correct in general: I'm going to need to go back and fix
up some input widths.  These should all be easier to find as at least
now the variable names all have the correct widths (it'll just be some
operations that don't).