Add scaffolding for safer code generation
The current code generator simply consists of a whole bunch of
printf() calls.  This is a bit nasty because it's very easy to break
the output in subtle ways.  For example, there's nothing that enforces
that function prototypes match between difference source files.

This aims to solve that problem by representing the artifacts that
need to be placed in the generated output files with some C++ clasess
that can then be parsed by multiple output formatters to actually
produce the required output code.