Subclass libflo::node to node
The general idea behind this is that I want to define my own node
types so I can have extra data tagged to them.  This extra data will
include things like the function that accesses them and will
(hopefully) allow for a somewhat type-safe way of managing this name
mangling.