Move (some) name mangling to node
The general idea here is that I only want to mangle these internal
names that refer to Chisel nodes, not those that refer to manually
named values (such as those specified when generating function
arguments).

This moves this particular sort of name mangling into node, as it's
really a Flo-specific mangling that's being performed here.