Handle ":" or "::" as a path seperator
It turns out that the C++ backend uses both "_" and "__" as path
seperators.  I have no idea why this happens, but I decided to match
the same behavior in the Flo backend.  This way I can simply "sed
s/:/_/g" to get the C++ name from a Flo name.

Essentially this takes the oddness of the C++ backend and duplicates
that complexity in Chisel as opposed to duplicating it here.