Be sure to emit all copied nodes to the header
If Flo like

  T0 = lit/1 1
  Design::reg = reg/1 T0

is emitted, then the old code wouldn't work because clock_hi would
copy Design__reg <= T0.  This new code makes sure that T0 is always
emitted to the header file, even when other rules don't require that.

This is slightly more efficient than how the current Chisel code does
it (with shadow values) because we can sometimes avoid emiting an
extra copy-to-header operation.