Add a copy of the Chisel JAR file
This allows me to run tests just from a Scala file.  This has
advantages and disadvantages: it allows me to test that I still
actually work with the Flo backend, but it also requires that the C++
backend for Chisel works.

The testing infastructure is kind of a mess...  I really need to
figure out what I'm doing here.