Fix race in tests named "test"
I had the line

  $ cat test.stdin | ./opt

which would both read from and overwrite test.stdin, causing a race
condition.  This only happened for tests named "test", which is why I
hadn't noticed it before.

The fix is to just copy and run

  $ cat test.stdin.copy | ./opt

I also remove a workaround for this overwriting, though the removed
workaround just helps me read stdin (not the computer).