Add support for "." demangling
The multi-word expansion code emits some signals that have a "." in
them.  These have some special semantics: they shouldn't be exported
(unless they should be for another reason), but the dot character
needs to be stripped so the C++ backend doesn't barf.

The new code supports this extension.  There's also a small test, but
it doesn't really do anything.