Call the header file "test.h", not "test.h++"
This matches (and therefor overwrites) the header file generated by
Chisel'l C++ backend.  This is super important because you really only
want to compile with one of the headers!  The old practice led to the
tendency to build with one of each header file (Chisel was still
generating an emulator for the Scala tests), which is all kinds of
bad...