Update to the latest Chisel
Note that this is a _huge_ commit: it touches all the tutgen tests
like usual, but also invalidates some of the cached-chisel tests and
requires an update to the header/compat generation code.

Unfortunately this really all has to be atomic so I guess there's not
anything else I can do.