Remove node::ptr_func()
This was a bit of nastiness from a long time ago: essentially I used
to handle C++ name demangling by providing a function that returns a
pointer to each dat_t in the header, and then another function to
set/get for each size of dat_t.  This was super nasty because I had a
special case of iterating through nodes of a particular width.

The new code is significantly simpler: now I just emit a get/set
function for each node.  This will make handling memories a bunch
easier because I won't have to have a memory type for libcodegen.