Add a DES test case
This is the first large test case I've ever run.  It's definately
bigger than any of the other tests, but it doesn't actually appear to
be any more complicated -- the test pretty much just does a whole
bunch of single-bit operations.

DES does, however, provide a pretty decent test for the VCD dumper.
It found a whole bunch of bugs related to designs that actually have
more than 3 levels in their module hierarchy, which is probably a good
thing.

It's also the first test that's big enough to potientally benchmark,
which could be interesting!