Run all tests with valgrind
Note that this is a bit of a hack: I don't actually verify that the
valgrind-rn outputs are correct, just that the regular outputs are
correct and that valgrind doesn't report any failures.

This should probably eventually get fixed but I'll need to clean up
the test harness (which will probably never happen :)).