Handle write-enable in WR nodes
I guess I forgot to handle write enable, which is now handled by a
read-modify-write cycle.  There isn't a test for this now, but
chisel_mem.bash fails if we diff against temp nodes, which will be
coming in a future Chisel release (that's how I found the bug, but my
current implementation is pretty hacky).