Don't output register inputs to VCDs in release
I'm not sure this is actually ever the correct thing to do, but for
now I'm just being conservative.  This used to export a _chisel_temps_
section with the inputs of registers but now doesn't.  These inputs
need to be in the header file but don't need to be in the VCD file,
which is why I think it's generally wrong.