Add a "--torture" option
I'm not sure what's going on with Torture.  It appears this may
actually be a Chisel bug, as I don't think these registers should end
up in the output circuit.

Regardless, I've just gone ahead and thrown together a version of
flo-llvm that works properly with flo-torture.  For now I think it's
good enough.