Implement LOG2
There's a test for this one.  Note that the current implementation is
pretty horrible (it uses an unrolled loop), but it's the best I can
come up with because LLVM doesn't appear to have a log2 intrinsic
(well, I could probably write a tighter loop, but who cares).

I'd consider the lack of a log2 intrinsic an LLVM bug...