Print a better error message when there's no class name
C++ needs a class name in order to build an emulator (and to wrap the
VCD file).  There's always been an abort() when flo-llvm is unable to
determine the class name, but this adds another line to give the user
a bit of a hint as to what to do.