Link LLVM IR directly on OSX
This one is a huge mess and might break: essentially the idea here is
that the LLVM compiler doesn't actually care to emit assembly code
that the assembler understands (at least on OSX).  The answer is to
rely on LLVM's integrated assembler to go from IR to object file in
one shot.

I don't understand why they don't consider this a bug...