Sanitize shadow copy handling
The previous shadow copy stuff would only work if there was already a
header value for the register's input signal.  While this used to be
guarnteed via a hack that made all register's inputs availiable in the
C++ header, it is now handled via a more generic shadow copy
mechanism, which will be used to deal with memory shadow copies.